% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lqmix.R
\name{coef.lqmix}
\alias{coef.lqmix}
\title{Print the estimated coefficients of an \code{lqmix} object}
\usage{
\method{coef}{lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
Return the estimated coefficients in the longitudinal data model obtained at convergence of the EM algorithm for a fitted model of \code{\link{class}} \code{\link{lqmix}}.
}
\description{
Print the estimated coefficients of a fitted model stored in an object of \code{\link{class}} \code{\link{lqmix}}.
}
