% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lqr.R
\name{coef.lqr}
\alias{coef.lqr}
\title{Print the estimated fixed coefficients of an \code{lqr} object}
\usage{
\method{coef}{lqr}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
Return the estimated coefficients obtained at convergence of the EM algorithm for a fitted model of \code{\link{class}} \code{lqr}
}
\description{
Print the estimated fixed coefficients of a fitted model of \code{\link{class}} \code{lqr}
}
