% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.seach_lqmix.R
\name{coef.search_lqmix}
\alias{coef.search_lqmix}
\title{Printn the estimated coefficients for the optimal fitted model}
\usage{
\method{coef}{search_lqmix}(object, ...)
}
\arguments{
\item{object}{a \code{search_lqmix} object}

\item{\dots}{not used}
}
\value{
Return the estimated coefficients obtained at convergence of the EM algorithm for the optimal model obtained at convergence of the EM algorithm for a fitted model of \code{\link{class}} \code{\link{search_lqmix}}.
}
\description{
Print the estimated coefficients of the optimal fitted model stored in an object of \code{\link{class}} \code{\link{search_lqmix}}.
}
