% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.lqr.R
\name{logLik.lqr}
\alias{logLik.lqr}
\title{Print the log-likelihood of an \code{lqr} object}
\usage{
\method{logLik}{lqr}(object, ...)
}
\arguments{
\item{object}{an \code{lqr} object}

\item{\dots}{not used}
}
\value{
Return an object of \code{\link{class}} \code{logLik} providing the log-likelihood for a fitted model of \code{\link{class}} \code{lqr}
}
\description{
Print the log-likelihood of a fitted model of \code{\link{class}} \code{lqr}
}
