% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.search_lqmix.R
\name{logLik.search_lqmix}
\alias{logLik.search_lqmix}
\title{Print the log-likelihood of the optimal model stored in a \code{search_lqmix} object}
\usage{
\method{logLik}{search_lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
Return an object of \code{\link{class}} \code{logLik} providing the log-likelihood value at convergence of the EM algorithm for a fitted model of \code{\link{class}} \code{\link{lqmix}}.
}
\description{
Print the log-likelihood of an optimal fitted model stored in an object of \code{\link{class}} \code{\link{search_lqmix}}.
}
