% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lqmix.R
\name{plot.lqmix}
\alias{plot.lqmix}
\title{Plots for \code{lqmix} objects}
\usage{
\method{plot}{lqmix}(x, ...)
}
\arguments{
\item{x}{an object of class \code{lqmix}}

\item{\dots}{not used}
}
\description{
Graphically display component and/or transition probabilities of a fitted model of \code{\link{class}} \code{\link{lqmix}}.
}
