% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lqr.R
\name{predict.lqr}
\alias{predict.lqr}
\title{Predictions from an \code{lqr} object}
\usage{
\method{predict}{lqr}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an \code{lqr} object}

\item{newdata}{an optional data frame in which to look for variables with which to predict. If omitted, the fitted values are produced}

\item{\dots}{not used}
}
\value{
A vector of predictions.
}
\description{
Returns the predicted values for an object of \code{\link{class}} \code{\link{lqr}}.
}
\details{
The function computes predictions for an object of \code{\link{class}} \code{\link{lqr}}.
}
