% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.search_lqmix.R
\name{predict.search_lqmix}
\alias{predict.search_lqmix}
\title{Predictions from the optimal fitted model}
\usage{
\method{predict}{search_lqmix}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a \code{search_lqmix} object}

\item{newdata}{an optional data frame in which to look for variables with which to predict. If omitted, the fitted values are produced}

\item{\dots}{not used}
}
\value{
A vector, a matrix, or an array of predictions, based on the optimal fitted model.
}
\description{
Returns the predicted values of the optimal fitted model stored in an object of \code{\link{class}} \code{\link{search_lqmix}}.
}
\details{
The function computes predictions for the optimal fitted model stored in an object of \code{\link{class}} \code{search_lqmix}.
If the optimal fitted model is based on TC, discrete, random coefficients only, a matrix of size \code{nsbjs x G} is given as output; if the optimal fitted model is based on TV, discrete, random coefficients only, a matrix of size \code{nObs x m}; if the optimal fitted model is based on both TC and TV, discrete, random coefficients, an array of size \code{nObs x G x m} is returned.
}
