% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lqmix.R
\name{print.lqmix}
\alias{print.lqmix}
\title{Print an \code{lqmix} Object}
\usage{
\method{print}{lqmix}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an \code{lqmix} object}

\item{digits}{a non-null value for digits specifies the mininum number of significant digits to be printed}

\item{...}{not used}
}
\value{
Return an \code{lqmix} object
}
\description{
Print an object of \code{\link{class}} \code{lqmix}
}
