% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lqr.R
\name{print.lqr}
\alias{print.lqr}
\title{Print an \code{lqr} object}
\usage{
\method{print}{lqr}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an \code{lqr} object}

\item{digits}{a non-null value for digits specifying the minimum number of significant digits to be printed}

\item{...}{not used}
}
\value{
Return an \code{\link{lqr}} object.
}
\description{
Print an object of \code{\link{class}} \code{\link{lqr}}.
}
