% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.search_lqmix.R
\name{print.search_lqmix}
\alias{print.search_lqmix}
\title{Print a \code{search_lqmix} object}
\usage{
\method{print}{search_lqmix}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{a \code{search_lqmix} object}

\item{digits}{a non-null value for digits specifying the minimum number of significant digits to be printed}

\item{...}{not used}
}
\value{
Return a \code{search_lqmix} object
}
\description{
Print an object of \code{\link{class}} \code{search_lqmix}
}
