% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.lqmix.R
\name{print.summary.lqmix}
\alias{print.summary.lqmix}
\title{Print the summary of an \code{lqmix} object}
\usage{
\method{print}{summary.lqmix}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{a summary of an \code{lqmix} object}

\item{digits}{a non-null value for digits specifying the minimum number of significant digits to be printed}

\item{...}{not used}
}
\value{
Return a summary of an \code{lqmix} object
}
\description{
Print the summary of an object of \code{\link{class}} \code{lqmix}
}
