% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.lqr.R
\name{print.summary.lqr}
\alias{print.summary.lqr}
\title{Print the Summary of an \code{lqr} Object}
\usage{
\method{print}{summary.lqr}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{a summary of an \code{lqr} object}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to be printed}

\item{...}{not used}
}
\value{
Return a summary of an \code{lqr} object
}
\description{
Print the summary of an an object of \code{\link{class}} \code{lqr}
}
