% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.search_lqmix.R
\name{print.summary.search_lqmix}
\alias{print.summary.search_lqmix}
\title{Print the summary of a \code{search_lqmix} object}
\usage{
\method{print}{summary.search_lqmix}(x, digits = max(3, getOption("digits") -
  3), ...)
}
\arguments{
\item{x}{a summary of a \code{search_lqmix} object}

\item{digits}{a non-null value for digits specifying the minimum number of significant digits to be printed}

\item{...}{not used}
}
\value{
Return a summary of a \code{\link{search_lqmix}} object.
}
\description{
Print the summary of an object of \code{\link{class}} \code{\link{search_lqmix}}.
}
