% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.lqmix.R
\name{residuals.lqmix}
\alias{residuals.lqmix}
\title{Residuals from an \code{lqmix} object}
\usage{
\method{residuals}{lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
A matrix or an array of of residuals, based on the estimated model.
}
\description{
Returns the residuals from a fitted \code{\link{lqmix}} object.
}
\details{
The function computes residuals for an object of class \code{lqmix}.
If the fitted model is based on TC, discrete, random coefficients only, a matrix of size \code{nsbjs x G} is given as output; if the fitted model is based on TV, discrete, random coefficients only, a matrix a matrix of size \code{nObs x m} is given as output is given as output; if the fitted model is based on both TC and TV, discrete, random coefficients, an array of size \code{nObs x G x m} is returned.
If the estimated model is based on TC, discrete, random coefficients only, a matrix is given as output. The number of columns corresponds to the estimated number of components (G).
If the estimated model is based on TV, discrete, random coefficients only, a matrix is given as output. The number of columns corresponds to the estimated number of states (m).
If the estimated model is based on both TC and TV, discrete, random coefficients, an array is given as output. The second and third dimensions correspond to the estimated number of components (G) and states (m), respectively.
}
