% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.lqr.R
\name{residuals.lqr}
\alias{residuals.lqr}
\title{Residuals from an \code{lqr} object}
\usage{
\method{residuals}{lqr}(object, ...)
}
\arguments{
\item{object}{an \code{lqr} object}

\item{\dots}{not used}
}
\value{
A vector of residuals.
}
\description{
Returns the residuals from a fitted \code{\link{lqr}} object.
}
\details{
The function computes residuals for an object of \code{\link{class}} \code{\link{lqr}}.
}
