% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.search_lqmix.R
\name{residuals.search_lqmix}
\alias{residuals.search_lqmix}
\title{Residuals from the optimal fitted model}
\usage{
\method{residuals}{search_lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{search_lqmix} object}

\item{\dots}{not used}
}
\value{
A vector, a matrix, or an array of of residuals, based on the estimated model.
}
\description{
Returns the residuals from the optimal fitted model stored in an object of \code{\link{class}} \code{\link{search_lqmix}}.
}
\details{
The function computes residuals for the optimal fitted model stored in an object of \code{class} \code{search_lqmix}.
If the optimal fitted model is based on TC, discrete, random coefficients only, a matrix of size \code{nsbjs x G} is given as output; if the optimal fitted model is based on TV, discrete, random coefficients only, a matrix a matrix of size \code{nObs x m} is given as output is given as output; if the optimal fitted model is based on both TC and TV, discrete, random coefficients, an array of size \code{nObs x G x m} is returned.
}
