% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varAL.R
\name{varAL}
\alias{varAL}
\title{Variance of Asymmetric Laplace random variables}
\usage{
varAL(sigma, qtl)
}
\arguments{
\item{sigma}{scale parameter}

\item{qtl}{skewness parameter}
}
\value{
Return the variance of Asymmetric Laplace random variables for given scale (\code{sigma}) and skewness (\code{qtl}) parameters.
}
\description{
Compute the variance for the asymmetric Laplace Distribution.
}
\references{
{
  \insertRef{ref:dal1}{lqmix}
}

{
  \insertRef{ref:dal2}{lqmix}
}

{
  \insertRef{ref:dal3}{lqmix}
}
}
