% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.lqmix.R
\name{vcov.lqmix}
\alias{vcov.lqmix}
\title{Calculate variance-covariance matrix for a fitted \code{lqmix} object}
\usage{
\method{vcov}{lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
A matrix of the estimated covariances between the parameter estimates in the linear predictor of the model.
}
\description{
Return the bootstrap variance-covariance matrix for the coefficients of a fitted \code{\link{lqmix}} object.
}
