% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.lqr.R
\name{vcov.lqr}
\alias{vcov.lqr}
\title{Calculate variance-covariance matrix for a fitted \code{lqr} object}
\usage{
\method{vcov}{lqr}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
A matrix of the estimated covariances between the parameter estimates in the linear predictor of the model.
}
\description{
Return the bootstrap variance-covariance matrix for the parameters of a fitted \code{\link{lqr}} object.
}
