% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.search_lqmix.R
\name{vcov.search_lqmix}
\alias{vcov.search_lqmix}
\title{Calculate variance-covariance matrix for the optimal fitted model}
\usage{
\method{vcov}{search_lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{search_lqmix} object}

\item{\dots}{not used}
}
\value{
A matrix of the estimated covariances between the main parameter estimates in the linear predictor of the model.
}
\description{
Return the bootstrap variance-covariance matrix of the parameters of the optimal fitted model stored in an object of \code{\link{class}} \code{\link{search_lqmix}}.
}
