% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lqmix.R
\name{coef.lqmix}
\alias{coef.lqmix}
\title{Print the estimated fixed coefficients of an \code{lqmix} object}
\usage{
\method{coef}{lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
Return the estimated fixed coefficients obtained at convergence of the EM algorithm for a fitted model of \code{\link{class}} \code{lqmix}
}
\description{
Print the estimated fixed coefficients of a fitted model of \code{\link{class}} \code{lqmix}
}
