% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lqr.R
\name{coef.lqr}
\alias{coef.lqr}
\title{Print the estimated coefficients of an \code{lqr} object}
\usage{
\method{coef}{lqr}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
Return the estimated coefficients for a fitted model of \code{\link{class}} \code{\link{lqr}}.
}
\description{
Print the estimated coefficients of a fitted model stored in an object of \code{\link{class}} \code{\link{lqr}}.
}
