% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.seach_lqmix.R
\name{coef.search_lqmix}
\alias{coef.search_lqmix}
\title{Print the estimated fixed coefficients of the optimal model stored in a \code{search_lqmix} object}
\usage{
\method{coef}{search_lqmix}(object, ...)
}
\arguments{
\item{object}{an \code{lqmix} object}

\item{\dots}{not used}
}
\value{
Return the estimated fixed coefficients obtained at convergence of the EM algorithm for the optimal model stored in an object of \code{\link{class}} \code{search_lqmix}
}
\description{
Print the estimated fixed coefficients of the optimal fitted model stored in an object of \code{\link{class}} \code{search_lqmix}
}
