% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{dal}
\alias{dal}
\title{Density of the Asymmetric Laplace Distribution}
\usage{
dal(y, mu = 0, sigma = 1, qtl = 0.5, log = FALSE)
}
\arguments{
\item{y}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{qtl}{skewness parameter}

\item{log}{logical; if TRUE, probabilities are log-transformed}
}
\value{
Return the density for the asymmetric Laplace distribution
}
\description{
Compute the density for the asymmetric Laplace distribution
}
