% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varAL.R
\name{varAL}
\alias{varAL}
\title{Variance of Asymmetric Laplace random variables}
\usage{
varAL(sigma, qtl)
}
\arguments{
\item{sigma}{scale parameter}

\item{qtl}{skewness parameter}
}
\value{
Return the variance of Asymmetric Laplace random variables for given scale (\code{sigma}) and skewness (\code{qtl}) parameters
}
\description{
Compute the variance for the asymmetric Laplace distribution
}
